add spawn sniper
