+++ .gitignore
@@ -2,3 +2,4 @@
 src/.DS_Store
 .DS_Store
 /.idea/
+book/book
