+++ assets/levels/lev01.json
@@ -248,7 +248,7 @@
                 {
                  "height":0,
                  "id":53,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
