+++ assets/levels/lev02.json
@@ -224,7 +224,7 @@
                 {
                  "height":0,
                  "id":56,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
