+++ assets/levels/lev03.json
@@ -224,7 +224,7 @@
                 {
                  "height":0,
                  "id":53,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
