+++ assets/levels/lev05.json
@@ -164,7 +164,7 @@
                 {
                  "height":0,
                  "id":53,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
