+++ assets/levels/lev06.json
@@ -200,7 +200,7 @@
                 {
                  "height":0,
                  "id":84,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
