+++ assets/levels/test_level.json
@@ -260,7 +260,7 @@
                 {
                  "height":0,
                  "id":147,
-                 "name":"seaweed",
+                 "name":"sniper",
                  "point":true,
                  "rotation":0,
                  "type":"",
@@ -272,7 +272,7 @@
                 {
                  "height":0,
                  "id":150,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
