+++ assets/levels/zyrafa01.json
@@ -219,7 +219,7 @@
                 {
                  "height":0,
                  "id":77,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
