+++ assets/levels/zyrafa02.json
@@ -159,7 +159,7 @@
                 {
                  "height":0,
                  "id":85,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
@@ -171,7 +171,7 @@
                 {
                  "height":0,
                  "id":86,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
@@ -183,7 +183,7 @@
                 {
                  "height":0,
                  "id":87,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
@@ -195,7 +195,7 @@
                 {
                  "height":0,
                  "id":88,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
