+++ assets/levels/zyrafa03.json
@@ -183,7 +183,7 @@
                 {
                  "height":0,
                  "id":87,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
@@ -195,7 +195,7 @@
                 {
                  "height":0,
                  "id":88,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
