+++ assets/levels/zyrafa04.json
@@ -171,7 +171,7 @@
                 {
                  "height":0,
                  "id":130,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
