+++ assets/levels/zyrafa05.json
@@ -183,7 +183,7 @@
                 {
                  "height":0,
                  "id":108,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
