+++ assets/map.json
@@ -284,7 +284,7 @@
                 {
                  "height":0,
                  "id":53,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
@@ -296,7 +296,7 @@
                 {
                  "height":0,
                  "id":54,
-                 "name":"muscet",
+                 "name":"musket",
                  "point":true,
                  "rotation":0,
                  "type":"",
