+++ book/src/simple_weapon.md
@@ -1,10 +1,28 @@
 # Simple Weapon
 
+In this section will walk through making a simple projectile weapon, a sniper rifle. The sniper rifle will be a variation of the musket weapon.
+
+## Planning
+
+The sniper rifle weapon will be very similar to the musket weapon, but with two distinct differences. The sniper rifle will have:
+
+- High shot speed
+- Low fire rate
+- Low ammunition
+
 ## Code
+
+Open the "FishFight" directory in your editor of choice
+
+### Create `sniper.rs` in `src/items`
+
+Create a new rust file in `src/items` called `your_new_item.rs`. I'll name the item file for this example `sniper.rs`.
+
+
 
 - `src/items.rs`: Add Item entry to ITEMS array
 - `src/main.rs`: Add resources to Resources?
-- `src/items`: Add file for weapon
+- `src/items`: Add file for weapon or add spawn in another weapon file
 
 ## Assets
